add dep graph for local install
